moving ConfirmDelete to StyledComponents
