+++ src/pages/TripPage.jsx
@@ -7,7 +7,7 @@ import Locations from "../components/Location/Locations"
 import Photos from "../components/Photos/Photos"
 import Memories from "../components/Memory/Memories"
 import Contributors from "../components/Contributors/Contributors"
-import ConfirmDelete from "../components/Utilities/ConfirmDelete"
+import ConfirmDelete from "../components/StyledComponents/ConfirmDelete"
 import StyledButton from "../components/StyledComponents/StyledButton"
 import { FaTrash, FaShare, FaPlaneDeparture } from "react-icons/fa6"
 import { useNavigate } from "react-router-dom"
